push: Decouple check and push parameters
Add two parameters specifically for the file presence checking, so both
the check and push processes have different set of parameters to define
number of goroutines and max number of file to process per request.
This is driven by the fact that these two actions differ and require
different params to adjust to proxying through the CF proxy.

Signed-off-by: Mike Sul <mike.sul@foundries.io>